Make behavioral utils shared
